extract chat sidebar
